+++ Web Art Gallery/WAG WebApp/Controllers/OrderController.cs
@@ -27,6 +27,11 @@ namespace WAG.WebApp.Controllers
 
         public IActionResult MakeOrder(int id, MakeOrderViewModel makeOrderViewModel)
         {
+            if (this.ArtisticWorkService.GetArtisticWorkById(id) == null)
+            {
+                return RedirectToAction("Categories", "ArtisticWork");
+            }
+
             makeOrderViewModel.ArtWorkId = id;
 
             return this.View(makeOrderViewModel);
@@ -35,11 +40,16 @@ namespace WAG.WebApp.Controllers
         [HttpPost]
         public IActionResult MakeOrder(MakeOrderViewModel makeOrderViewModel)
         {
+            if (this.ArtisticWorkService.GetArtisticWorkById(makeOrderViewModel.ArtWorkId) == null)
+            {
+                return RedirectToAction("Categories", "ArtisticWork");
+            }
+
             var currUser = this.UserAccountService.GetCurrentUser(HttpContext);
 
             this.OrderService.SaveOrder(currUser, makeOrderViewModel);
             
-            return RedirectToAction("Success", "Home", new { area = "" });
+            return RedirectToAction("Success", "Home");
         }
 
         public IActionResult MyOrders(MyOrdersViewModel myOrdersViewModel)
@@ -58,12 +68,17 @@ namespace WAG.WebApp.Controllers
 
             var order = this.OrderService.GetOrderById(id);
 
+            if (order == null)
+            {
+                return RedirectToAction("MyOrders", "Order");
+            }
+
             if (order.WAGUserId == currUser.Id)
             {
                 this.OrderService.DeleteOrder(order.Id);
             }
 
-            return RedirectToAction("Success", "Home", new { area = "" });
+            return RedirectToAction("Success", "Home");
         }
 
         public IActionResult MakeSpecialOrder()
@@ -83,12 +98,17 @@ namespace WAG.WebApp.Controllers
 
             this.OrderService.SaveSpecialOrder(currUser, makeSpecialOrderViewModel);
 
-            return RedirectToAction("Success", "Home", new { area = "" });
+            return RedirectToAction("Success", "Home");
         }
 
         public IActionResult OrderDetails(int id, OrderDetailsViewModel orderDetailsViewModel)
         {
             orderDetailsViewModel.Order = this.OrderService.GetOrderById(id);
+
+            if (orderDetailsViewModel.Order == null)
+            {
+                return RedirectToAction("MyOrders", "Order");
+            }
 
             if (orderDetailsViewModel.Order.ArtisticWorkId != null )
             {
